requirements.txt
git+https://github.com/gresas/rest_clients#egg=rest_clients
dependency_injector<5
PyYAML<7
pymongo<5
requests<3
redis<8
urllib3<3
